backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
ultralytics==8.0.196
Pillow==10.0.0
numpy==1.24.3 